requirements.txt
python-dotenv>=1.0.1
pyyaml>=6.0.2
tqdm>=4.66.4
pandas>=2.2.2
openai>=1.40.0            
dashscope>=1.19.3         
langchain>=0.2.11
langchain-openai>=0.1.22  
scikit-learn>=1.4         
